moved fastword function to pkg

changed fastforward


squash commits


changed ff cmd

fixed ff cmd


fixed ff_test


removed unused func


fixing tests


revert to the getFfOptions


removed sut_test:getRootOptions


Update cmd/krel/cmd/ff_test.go
