+++ cmd/krel/cmd/ff.go
@@ -21,23 +21,13 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
+	"k8s.io/release/pkg/fastforward"
 	kgit "k8s.io/release/pkg/git"
-	"k8s.io/release/pkg/util"
 )
 
-type ffOptions struct {
-	repoPath       string
-	branch         string
-	mainRef        string
-	nonInteractive bool
-	cleanup        bool
-}
-
-var ffOpts = &ffOptions{}
+var ffOpts = &fastforward.Options{}
 
 // ffCmd represents the base command when called without any subcommands
 var ffCmd = &cobra.Command{
@@ -62,172 +52,16 @@ as real push if the '--nomock' flag is specified.
 	SilenceUsage:  true,
 	SilenceErrors: true,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		return runFf(ffOpts, rootOpts)
+		ffOpts.NoMock = rootOpts.nomock
+		return fastforward.Run(ffOpts)
 	},
 }
 
-const pushUpstreamQuestion = `Are you ready to push the local branch fast-forward changes upstream?
-Please only answer after you have validated the changes.`
-
 func init() {
-	ffCmd.PersistentFlags().StringVar(&ffOpts.repoPath, "repo", filepath.Join(os.TempDir(), "k8s"), "the local path to the repository to be used")
-	ffCmd.PersistentFlags().StringVar(&ffOpts.branch, "branch", "", "branch")
-	ffCmd.PersistentFlags().StringVar(&ffOpts.mainRef, "ref", kgit.Remotify(kgit.DefaultBranch), "ref on the main branch")
-	ffCmd.PersistentFlags().BoolVar(&ffOpts.cleanup, "cleanup", false, "cleanup the repository after the run")
+	ffCmd.PersistentFlags().StringVar(&ffOpts.RepoPath, "repo", filepath.Join(os.TempDir(), "k8s"), "the local path to the repository to be used")
+	ffCmd.PersistentFlags().StringVar(&ffOpts.Branch, "branch", "", "branch")
+	ffCmd.PersistentFlags().StringVar(&ffOpts.MainRef, "ref", kgit.Remotify(kgit.DefaultBranch), "ref on the main branch")
+	ffCmd.PersistentFlags().BoolVar(&ffOpts.Cleanup, "cleanup", false, "cleanup the repository after the run")
 
 	rootCmd.AddCommand(ffCmd)
-}
-
-func runFf(opts *ffOptions, rootOpts *rootOptions) error {
-	branch := opts.branch
-	if branch == "" {
-		return errors.New("please specify valid release branch")
-	}
-
-	logrus.Infof("Preparing to fast-forward %s onto the %s branch", opts.mainRef, branch)
-	repo, err := kgit.CloneOrOpenDefaultGitHubRepoSSH(opts.repoPath)
-	if err != nil {
-		return err
-	}
-
-	if !rootOpts.nomock {
-		logrus.Info("Using dry mode, which does not modify any remote content")
-		repo.SetDry()
-	}
-
-	logrus.Infof("Checking if %q is a release branch", branch)
-	if isReleaseBranch := kgit.IsReleaseBranch(branch); !isReleaseBranch {
-		return errors.Errorf("%s is not a release branch", branch)
-	}
-
-	logrus.Info("Checking if branch is available on the default remote")
-	branchExists, err := repo.HasRemoteBranch(branch)
-	if err != nil {
-		return errors.Wrap(err, "checking if branch exists on the default remote")
-	}
-	if !branchExists {
-		return errors.New("branch does not exist on the default remote")
-	}
-
-	if ffOpts.cleanup {
-		defer repo.Cleanup() // nolint: errcheck
-	} else {
-		// Restore the currently checked out branch afterwards
-		currentBranch, err := repo.CurrentBranch()
-		if err != nil {
-			return errors.Wrap(err, "unable to retrieve current branch")
-		}
-		defer func() {
-			if err := repo.Checkout(currentBranch); err != nil {
-				logrus.Errorf("Unable to restore branch %s: %v", currentBranch, err)
-			}
-		}()
-	}
-
-	logrus.Info("Checking out release branch")
-	if err := repo.Checkout(branch); err != nil {
-		return errors.Wrapf(err, "checking out branch %s", branch)
-	}
-
-	logrus.Infof("Finding merge base between %q and %q", kgit.DefaultBranch, branch)
-	mergeBase, err := repo.MergeBase(kgit.DefaultBranch, branch)
-	if err != nil {
-		return err
-	}
-
-	// Verify the tags
-	mainTag, err := repo.Describe(
-		kgit.NewDescribeOptions().
-			WithRevision(kgit.Remotify(kgit.DefaultBranch)).
-			WithAbbrev(0).
-			WithTags(),
-	)
-	if err != nil {
-		return err
-	}
-	mergeBaseTag, err := repo.Describe(
-		kgit.NewDescribeOptions().
-			WithRevision(mergeBase).
-			WithAbbrev(0).
-			WithTags(),
-	)
-	if err != nil {
-		return err
-	}
-	logrus.Infof("Merge base tag is: %s", mergeBaseTag)
-
-	if mainTag != mergeBaseTag {
-		return errors.Errorf(
-			"unable to fast forward: tag %q does not match %q",
-			mainTag, mergeBaseTag,
-		)
-	}
-	logrus.Infof("Verified that the latest tag on the main branch is the same as the merge base tag")
-
-	releaseRev, err := repo.Head()
-	if err != nil {
-		return err
-	}
-	logrus.Infof("Latest release branch revision is %s", releaseRev)
-
-	logrus.Info("Merging main branch changes into release branch")
-	if err := repo.Merge(opts.mainRef); err != nil {
-		return err
-	}
-
-	headRev, err := repo.Head()
-	if err != nil {
-		return err
-	}
-
-	prepushMessage(repo.Dir(), branch, opts.mainRef, releaseRev, headRev)
-
-	pushUpstream := false
-	if opts.nonInteractive {
-		pushUpstream = true
-	} else {
-		_, pushUpstream, err = util.Ask(pushUpstreamQuestion, "yes", 3)
-		if err != nil {
-			return err
-		}
-	}
-
-	if pushUpstream {
-		logrus.Infof("Pushing %s branch", branch)
-		if err := repo.Push(branch); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func prepushMessage(gitRoot, branch, ref, releaseRev, headRev string) {
-	fmt.Printf(`Go look around in %s to make sure things look okay before pushing…
-
-Check for files left uncommitted using:
-
-	git status -s
-
-Validate the fast-forward commit using:
-
-	git show
-
-Validate the changes pulled in from main branch using:
-
-	git log %s..%s
-
-Once the branch fast-forward is complete, the diff will be available after push at:
-
-	https://github.com/%s/%s/compare/%s...%s
-
-`,
-		gitRoot,
-		kgit.Remotify(branch),
-		ref,
-		kgit.DefaultGithubOrg,
-		kgit.DefaultGithubRepo,
-		releaseRev[:11],
-		headRev[:11],
-	)
 }
